[workflow] Create ccpp.yml
An important part of CICD includes automated testing.
This commits enables free tier GitHub `Action` to evaluate and
run tests upon pushes and open PRs.
It also creates daily cron job to run tests and notify failures.
This workflow is important because it helps with the trivial
job of making sure new commits to the Repo always pass every test.