Iterative instead of recursive implementation of relocation
Both recursive and iterative implementations share the most of the
issues. However, all the implementations available to me use the
iterative version.

Iterative implementation maintains the relocation chain in a separate
data structure (stack) until its length does not exceed the hard bound
on relocation count. Once that happen, insert procedure
fails. Otherwise, the relocation chain is committed to the storage of
the cuckoo filter.